algorithms/PWC-Net/UV_files/flowiz/requirements.txt
tqdm==4.59.0
matplotlib==3.3.4
numpy
Eel==0.14.0
Pillow==8.3.2
